Add files via upload
Day 53 of Data Science journey(Python)

🌴 Exception Handling

✅ UNIT-14_ASSIGNMENT(COMPLETED)

📍 Choose the correct option for the following questions 

1) What is the purpose of the try and except keywords in Python? 
A) To handle errors and exceptions in a program. 
B) To perform conditional operations. 
C) To create user-defined functions. 
D) To import modules. 

➡️ Answer : A

2) What happens if an exception is raised in the try block? 
A) The except block will be executed. 
B) The program will continue to run. 
C) The program will stop executing. 
D) None of the above. 

➡️Answer - A

3) What happens if no exception is raised in the try block? 
A) The except block will be executed. 
B) The program will continue to run. 
C) The program will stop executing. 
D) None of the above.
 
➡️Answer -B

4) What is the use of the finally keyword in exception handling? 
A) To handle exceptions and errors. 
B) To specify a block of code that will be executed regardless of whether an 
exception was raised or not. 
C) To create user-defined functions. 
D) None of the above 
➡️Answer B

5) What is a custom exception in Python? 
A) A built-in exception 
B) A user-defined exception 
C) Both A and B 
D) none of the above

➡️Answer - B

6) can you create a custom exception in Python? 
A) By using the raise statement 
B) By using the assert statement 
C) Both A and B 
D) none of the above

➡️Answer - A